requirements.txt
autopep8==1.7.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paho-mqtt==2.1.0
pycodestyle==2.9.1
pygame==2.1.2
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dotenv==1.1.0
requests==2.29.0
shiboken6==6.9.0
toml==0.10.2
urllib3==1.26.15
Werkzeug==2.3.2
